Put Raid into own file.
